agents/a2a-translator-gateway/pyproject.toml
[project]
name = "a2a-translator"
version = "0.1.0"
description = "A2A Translator agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "openai>=1.57.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "langdetect>=1.0.9",
    "googletrans>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"